refactor: Refactor inventory report
